Refactored code to prepare for bare VCD 5.5 support
